refactor(assets): SVG import 경로 수정
