+++ src/page/forum-post/assets/CategoryArrow.tsx
@@ -1,4 +1,4 @@
-import SVGAttribute from "@/global/types/SVGAttribute";
+import SVGAttribute from "@/global/types/SVGAttribute.type";
 import React from "react";
 
 const CategoryArrow = ({ width, height, pointable }: SVGAttribute) => {
