+++ src/page/forum-post/assets/CommentIcon.tsx
@@ -1,4 +1,4 @@
-import SVGAttribute from "@/global/types/SVGAttribute";
+import SVGAttribute from "@/global/types/SVGAttribute.type";
 import React from "react";
 
 const CommentIcon = ({ width, height, pointable }: SVGAttribute) => {
