+++ src/page/forum-post/assets/Like.tsx
@@ -1,4 +1,4 @@
-import SVGAttribute from "@/global/types/SVGAttribute";
+import SVGAttribute from "@/global/types/SVGAttribute.type";
 import React from "react";
 
 const Like = ({ width, height, pointable }: SVGAttribute) => {
