+++ src/page/forum/assets/LikeLogo.tsx
@@ -1,4 +1,4 @@
-import SVGAttribute from "@/global/types/SVGAttribute";
+import SVGAttribute from "@/global/types/SVGAttribute.type";
 import React from "react";
 
 const LikeLogo = ({ width, height, pointable }: SVGAttribute) => {
